+++ packages/sampler/stories/qa-stories/Scroller.js
@@ -61,8 +61,7 @@ storiesOf('Scroller')
 						<Button key={i + 1}>
 							Button {i + 1}
 						</Button>
-					)
-				)}
+					))}
 				</div>
 			</Scroller>
 		)