+++ packages/spotlight/src/tests/container-specs.js
@@ -434,6 +434,38 @@ describe('container', () => {
 			}
 		));
 
+		it('should return the default spottable element when enterTo is "default-element" and defaultElement contains an array of selectors', testScenario(
+			scenarios.containerWithDefaultAndLastFocused,
+			() => {
+				configureContainer('container', {
+					enterTo: 'default-element',
+					defaultElement: ['.does-not-exist', '.spottable-default']
+				});
+
+				const expected = 'spottableDefault';
+				const actual = getContainerFocusTarget('container').id;
+
+				expect(actual).to.equal(expected);
+			}
+		));
+
+		// FIXME: This is testing a previously supported feature (setting a node as defaultElement)
+		// which was never documented and should be removed in a future release.
+		it('should return the default spottable element when enterTo is "default-element" and defaultElement contains an array of selectors wiht a node reference', testScenario(
+			scenarios.containerWithDefaultAndLastFocused,
+			(root) => {
+				configureContainer('container', {
+					enterTo: 'default-element',
+					defaultElement: [root.querySelector('#lastFocused'), '.spottable-default']
+				});
+
+				const expected = 'lastFocused';
+				const actual = getContainerFocusTarget('container').id;
+
+				expect(actual).to.equal(expected);
+			}
+		));
+
 		it('should return the first spottable element when enterTo is "default-element" but defaultElement is not configured', testScenario(
 			scenarios.containerWithDefaultAndLastFocused,
 			() => {