+++ packages/ui/CHANGELOG.md
@@ -12,6 +12,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
+- `ui/Cancelable` warning for string type cancel handler
 
 ## [1.4.1] - 2017-07-05